+++ src/rust/ide/view/graph-editor/src/builtin/visualization/java_script/table_view/style_light.css
@@ -1,0 +1,30 @@
+table {
+    font-family: sans-serif;
+    font-size: 12px;
+}
+
+td {
+    color: rgba(0, 0, 0, 0.9);
+    padding: 0;
+}
+
+td.plaintext,
+th {
+    padding: 5px;
+}
+
+th,
+td {
+    border: 1px solid transparent;
+    background-clip: padding-box;
+}
+
+th {
+    color: rgba(0, 0, 0, 0.7);
+    font-weight: 400;
+}
+
+td,
+th {
+    background-color: rgba(0, 0, 0, 0.03);
+}
